% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMSVmetrics.R
\name{estimateMSVmetrics}
\alias{estimateMSVmetrics}
\title{Estimates the GPD and SPOmulti-source variability metrics and the corresponding source projection vertices from a matrix of probability distributions of different data sources}
\usage{
estimateMSVmetrics(msvMetrics, nBySource, idSource)
}
\arguments{
\item{probabilities}{m-by-n matrix containing the probability mass of n data sources on m distribution bins}
}
\value{
A list containing the following results:
	GPD       	the value of the Global Probabilistic Deviation metric, where 0 means equal distributions and 1 means non-overlapping distributions
	SPOs      	the values of Source Probabilistic Outlyingness for each data source, where 0 means equal to central tendency and 1 completely non-overlapping
	Vertices	a n-by-(n-1) matrix containing the coordinates of each data source in the projected probabilistic space conserving their dissimilarities, e.g, for 3D projections the 3 first columns can be used
}
\description{
Estimates the GPD and SPOmulti-source variability metrics and the corresponding source projection vertices from a matrix of probability distributions of different data sources
}
